app/pyproject.toml
[tool.poetry]
name = "3grit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
aiohappyeyeballs = "2.4.4"
aiohttp = "3.11.10"
aiosignal = "1.3.2"
annotated-types = "0.7.0"
anyio = "4.5.2"
async-fastapi-jwt-auth = "0.6.6"
asyncpg = "0.30.0"
attrs = "24.3.0"
beautifulsoup4 = "4.12.3"
boto3 = "1.34.64"
botocore = "1.34.162"
bs4 = "0.0.2"
certifi = "2024.12.14"
charset-normalizer = "3.4.0"
click = "8.1.7"
dependency-injector = "4.44.0"
distro = "1.9.0"
dnspython = "2.6.1"
email-validator = "2.2.0"
fastapi = "0.115.4"
frozenlist = "1.5.0"
gunicorn = "23.0.0"
h11 = "0.14.0"
httpcore = "1.0.7"
httptools = "0.6.4"
httpx = "0.28.1"
idna = "3.10"
jiter = "0.8.2"
jmespath = "1.0.1"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
langchain = "0.3.7"
langchain-core = "0.3.25"
langchain-openai = "0.2.5"
langchain-text-splitters = "0.3.3"
langsmith = "0.1.147"
lxml = "5.3.0"
multidict = "6.1.0"
mysql-connector-python = "8.4.0"
numpy = "1.26.4"
openai = "1.53.0"
orjson = "3.10.12"
packaging = "24.2"
pillow = "10.2.0"
pinecone-client = "5.0.1"
pinecone-plugin-inference = "1.1.0"
pinecone-plugin-interface = "0.0.7"
propcache = "0.2.1"
py-ulid = "1.0.3"
pydantic = {extras = ["email"], version = "^2.10.4"}
pydantic-settings = "2.7.0"
pydantic-core = "2.27.2"
pyjwt = "2.10.1"
pymongo = "3.11.3"
python-dateutil = "2.9.0.post0"
python-docx = "0.8.11"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
pytz = "2024.2"
pyyaml = "6.0.2"
redis = "5.0.1"
regex = "2024.11.6"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
s3transfer = "0.10.4"
six = "1.17.0"
sniffio = "1.3.1"
soupsieve = "2.6"
sqlalchemy = "2.0.36"
starlette = "0.41.2"
tenacity = "9.0.0"
tiktoken = "0.8.0"
tqdm = "4.67.1"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
uuid = "1.30"
uvicorn = "0.32.0"
uvloop = "0.21.0"
watchfiles = "0.24.0"
websockets = "13.1"
yarl = "1.18.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
